Saving contract_subjects #144
+refactor
