+++ gpn/gpn.py
@@ -1,8 +1,3 @@
-from analyser.hyperparams import HyperParameters
-from analyser.structures import OrgStructuralLevel, legal_entity_types_as_db_json
-from analyser.text_tools import compare_masked_strings
-from integration.db import get_mongodb_connection
-
 data = {
   "Subsidiary": [
     {
@@ -674,49 +669,6 @@ data = {
 subsidiaries = data['Subsidiary']
 
 
-def all_do_names():
-  for s in subsidiaries:
-    for alias in s['aliases'] + [s['_id']]:
-      yield alias
 
 
-def estimate_subsidiary_name_match_min_jaro_similarity():
-  top_similarity = 0
 
-  for name1 in all_do_names():
-    for name2 in all_do_names():
-      name1 = name1.replace('»', '').replace('«', '')
-      name2 = name2.replace('»', '').replace('«', '')
-
-      if name1.lower() != name2.lower():
-
-        similarity = compare_masked_strings(name1, name2, [])
-        if similarity > top_similarity:
-          top_similarity = similarity
-          print(top_similarity, name1, name2)
-
-  return top_similarity
-
-
-def update_db_dictionaries():
-  db = get_mongodb_connection()
-
-  coll = db["subsidiaries"]
-  coll.delete_many({})
-  coll.insert_many(subsidiaries)
-
-  coll = db["orgStructuralLevel"]
-  coll.delete_many({})
-  coll.insert_many(OrgStructuralLevel.as_db_json())
-
-  coll = db["legalEntityTypes"]
-  coll.delete_many({})
-  coll.insert_many(legal_entity_types_as_db_json())
-
-
-HyperParameters.subsidiary_name_match_min_jaro_similarity = estimate_subsidiary_name_match_min_jaro_similarity()
-print('HyperParameters.subsidiary_name_match_min_jaro_similarity',
-      HyperParameters.subsidiary_name_match_min_jaro_similarity)
-
-if __name__ == '__main__':
-  update_db_dictionaries()
